+++ src/options.js
@@ -92,13 +92,14 @@ module.exports = optionator({
             alias: "x",
             type: "String",
             default: "all",
+            dependsOn: ["fix"],
             description: "List of rule identifiers to fix (if not specified, all available fixes will be applied)"
         },
         {
             heading: "Ignoring files"
         },
         {
-            option: "ignore-pattern",
+            option: "ignorepattern",
             alias: "i",
             type: "String",
             description: "Comma-separated list of Ant-style file patterns specifying files that must be ignored. Default: none",
@@ -140,6 +141,7 @@ module.exports = optionator({
     mutuallyExclusive: [
         ["files", "source", "codenarcargs", "help"],
         [["path", "files"], "source"],
-        ["failonerror", "failonwarning", "failoninfo"]
+        ["failonerror", "failonwarning", "failoninfo"],
+        ["codenarcargs", ["failonerror", "failonwarning", "failoninfo", "path", "files", "source", "fix", "fixrules"]]
     ]
 });