+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Capability to call NpmGroovyLint with options as object, not only command line arguments
 - New option "source", allowing to call NpmGroovyLint with the groovy code as a string , not only path & files pattern
+- Run lint again after fix to get updated error log
 
 ## [2.0.1] - 2020-02-21